+++ gr/gr_modem.h
@@ -115,7 +115,7 @@ public slots:
     void setTxPower(float value);
     void setBbGain(int value);
     void setSquelch(int value);
-    void setRxSensitivity(float value);
+    void setRxSensitivity(double value);
     void setRxCTCSS(float value);
     void setTxCTCSS(float value);
     void enableGUIConst(bool value);
